requirements.txt
discord.py>=2.6,<3
aiohttp>=3.9,<4
Pillow>=10.2.0,<11
